# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Spec for footbal team management",
        "title": "API Football Team Management",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "afrizalsebastian",
            "email": "sebastiangurning@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/players": {
            "get": {
                "description": "Get List Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Players"
                ],
                "summary": "Get List Player",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-array_dto_GetListPlayerItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/players/{playerId}": {
            "get": {
                "description": "Get Player Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Players"
                ],
                "summary": "Get Player Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player Id",
                        "name": "playerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_GetPlayerDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Player Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Players"
                ],
                "summary": "Update Player Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player Id",
                        "name": "playerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update player request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_GetPlayerDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Players"
                ],
                "summary": "Delete Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player Id",
                        "name": "playerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorsDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-array_dto_GetListPlayerItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetListPlayerItem"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-dto_GetPlayerDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GetPlayerDetailResponse"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetListPlayerItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/dto.PlayerPosition"
                },
                "team": {
                    "$ref": "#/definitions/dto.ListPlayerItemTeam"
                }
            }
        },
...
        "dto.GetPlayerDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "goal_count": {
                    "type": "integer"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/dto.PlayerPosition"
                },
                "team": {
                    "$ref": "#/definitions/dto.ListPlayerItemTeam"
                },
                "team_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "dto.ListPlayerItemTeam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.PlayerPosition": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePlayerRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Please input token with format: Bearer __TOKEN__",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}